requirements.txt
pygame-ce==2.5.6
typing-extensions==4.8.0
pytest==7.4.0
pytest-cov==4.1.0
pyinstaller==6.3.0
